LPS-74742 Change method signature to accept a long size

+++ modules/apps/adaptive-media/adaptive-media-image-service/src/main/java/com/liferay/adaptive/media/image/service/impl/AMImageEntryLocalServiceImpl.java
@@ -83,7 +83,7 @@ public class AMImageEntryLocalServiceImpl
 	public AMImageEntry addAMImageEntry(
 			AMImageConfigurationEntry amImageConfigurationEntry,
 			FileVersion fileVersion, int width, int height,
-			InputStream inputStream, int size)
+			InputStream inputStream, long size)
 		throws PortalException {
 
 		_checkDuplicates(
